added modules camera, shader
