+++ src/gl_bindings.rs
@@ -1,0 +1,2 @@
+include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
+
